neo4j_ingestion_service/requirements.txt
# Development/Testing Dependencies
pytest>=7.0.0
pytest-mock>=3.0.0
neo4j>=4.4.0 # Need driver for verification framework too

# Core Dependencies
grpcio
grpcio-tools
protobuf
neo4j
psycopg2-binary # For PostgreSQL connection
python-dotenv # For loading .env file

grpcio-tools>=1.48.0 # For generating code during build
grpcio-health-checking>=1.48.0 # For health checks
